Python/Sync: move get_request_error_class to a shared location
